Refactor and document `aws' module.
