+++ aws/external_test.go
@@ -1,7 +1,6 @@
 package aws
 
 import (
-	"math/rand"
 	"testing"
 )
